+++ omod/src/main/webapp/fragments/overview/orderedTests.gsp
@@ -250,17 +250,19 @@ recentVisitFactory.fetchVisitEncounterObs(visitId).then(function(data) {
     			}
   		};
 
-  		\$scope.closeAlert = function(index) {
-    			\$scope.alerts.splice(index, 1);
-    			\$scope.errortext = "";
-    			var deleteurl = "/" + OPENMRS_CONTEXT_PATH + "/ws/rest/v1/obs/" + \$scope.respuuid[index] + "?purge=true";
-    			\$scope.respuuid.splice(index, 1);
-			\$http.delete(deleteurl).then(function(response){
-				\$scope.statuscode = "Success";
-			}, function(response){
-				\$scope.statuscode = "Failed to delete Obs";
-			});
-  		};  
+	  		\$scope.closeAlert = function(index) {
+	  			if (\$scope.visitStatus) {
+		    		\$scope.alerts.splice(index, 1);
+		    		\$scope.errortext = "";
+		    		var deleteurl = "/" + OPENMRS_CONTEXT_PATH + "/ws/rest/v1/obs/" + \$scope.respuuid[index] + "?purge=true";
+		    		\$scope.respuuid.splice(index, 1);
+					\$http.delete(deleteurl).then(function(response){
+						\$scope.statuscode = "Success";
+					}, function(response){
+						\$scope.statuscode = "Failed to delete Obs";
+					});
+				}
+	  		};  
         });
   }, 2000);
